removed chart from item-table
